Refactor and add to branches
